\name{llbayesireg-package}
\alias{llbayesireg-package}
\docType{package}
\title{The L-Logistic Bayesian Regression}
\source{
The L-Losgistic distribution was introduced by Tadikamalla and Johnson (1982), which refer to this distribution as Logit-Logistic
distribution. Here, we have a new parameterization of the Logit-Logistic with the median as a parameter.
}
\description{Bayesian L-Logistic regression package, <URL:\url{https://www.imstat.org/wp-content/uploads/2018/05/BJPS397.pdf}>.
}
\details{Package to estimate an L-Logistic regression model with median and precision regression structures, diagnostics and HPD.
\tabular{ll}{
Package: \tab llbayesireg\cr
Type: \tab Package\cr
Version: \tab 0.1.0\cr
Date: \tab 2019-03-06\cr
License: \tab GPL-3\cr
LazyLoad: \tab yes\cr
}
}
\author{
 Sara Alexandre Fonsêca \email{saralexandre@alu.ufc.br},
 Rosineide Fernando da Paz \email{rfpaz2@gmail.com},
 Jorge Luís Bazán
}
\references{
Paz, R.F., Balakrishnan, N and Bazán, J.L. (2018). L-Logistic Distribution: Properties, Inference and an Application to Study Poverty and Inequality in Brazil.
}
\keyword{package}
